# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "WiGO PRO — API mobile",
        "version": "1.0.0",
        "description": "Contrat entre le back-office Laravel et l'application mobile **WiGO PRO**\n(chauffeurs Yango en Côte d'Ivoire, opéré par At Confort Plus).\n\n- Montants en **entiers FCFA** (XOF).\n- Dates au format **ISO 8601 UTC**.\n- Identifiants **ULID**.\n- Messages d'erreur **en français**, prêts à afficher.\n- Enveloppe unique : succès `{ message, data }`, erreur\n  `{ message, errors }`. Le code HTTP porte le statut, jamais le corps.\n- Pagination **curseur** : `meta.next_cursor` et `links.next`,\n  `per_page` ≤ 50.\n- Authentification par **jeton Sanctum** (habilitation `mobile:*`) obtenu\n  via `POST /auth/otp/verify`.\n\nMessagerie en direct (WebSocket) : voir [`/docs/api/guides/realtime`](/docs/api/guides/realtime)."
    },
    "servers": [
        {
            "url": "{origin}/api/v1",
            "description": "Serveur courant",
            "variables": {
                "origin": {
                    "default": "http://localhost:8000",
                    "description": "Origine de l'instance (schéma, domaine, port)."
                }
            }
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
...
    "paths": {
        "/auth/otp/request": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "v1.auth.otp.request",
                "description": "Envoie un code à 6 chiffres (valable 5 minutes) par SMS ou WhatsApp. Le\nconducteur doit être pré-enregistré par le back-office : l'application\nmobile ne crée pas de compte.\n\nLimité à 3 envois par tranche de 10 minutes et par numéro.\n\nEn développement et en test (`WIGO_OTP_EXPOSE_CODE`), la réponse contient\nen plus le champ `code` : il est toujours absent en production.",
                "summary": "Demander un code OTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OtpRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "examples": [
                                                "Un code de connexion vous a été envoyé. Il expire dans 5 minutes."
                                            ]
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "enum": [
                                                        "sms",
                                                        "whatsapp"
                                                    ]
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "examples": [
                                                        "2026-08-29T09:47:11Z"
                                                    ]
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "description": "Uniquement en développement et en recette\n(`WIGO_OTP_EXPOSE_CODE`) : absent en production.",
                                                    "examples": [
                                                        "482913"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "channel",
                                                "expires_at"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
...
        "/support/conversation": {
            "get": {
                "tags": [
                    "Support"
                ],
                "operationId": "v1.support.conversation",
                "description": "L'état de la conversation : aperçu du dernier message et nombre de\nmessages non lus, de quoi alimenter la pastille de l'écran d'accueil.\n\nLa conversation est créée à la volée si le conducteur n'a jamais écrit,\npour que l'application n'ait pas de cas particulier à traiter.",
                "summary": "Fil du conducteur",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Message prêt à afficher, en français."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ConversationResource"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview.",
                                    "examples": [
                                        "Authentification requise."
                                    ]
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview.",
                                    "examples": [
                                        "Les données envoyées sont invalides."
                                    ]
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "examples": [
                                                "Le champ téléphone doit être un numéro de téléphone valide."
                                            ]
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "ConversationResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "examples": [
                            "01HZY3M4N5P6Q7R8S9T0V1W2X3"
                        ]
                    },
                    "unread_count": {
                        "type": "integer",
                        "description": "Messages non lus par le conducteur.",
                        "examples": [
                            2
                        ]
                    },
                    "last_message_sender_type": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Auteur du dernier message. `null` s'il est système.",
                        "enum": [
                            "user",
                            "driver",
                            null
                        ]
                    },
                    "last_message_preview": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "examples": [
                            "Nous avons bien reçu votre demande."
                        ]
                    },
                    "last_message_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "examples": [
                            "2026-08-29T09:42:11Z"
                        ]
                    },
                    "driver_read_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "examples": [
                            "2026-08-29T09:45:03Z"
                        ]
                    }
                },
                "required": [
                    "id",
                    "unread_count",
                    "last_message_sender_type",
                    "last_message_preview",
                    "last_message_at",
                    "driver_read_at"
                ],
                "title": "ConversationResource"
            },
...
            "OtpChannel": {
                "type": "string",
                "enum": [
                    "whatsapp",
                    "sms"
                ],
                "title": "OtpChannel"
            },
            "OtpRequestRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string",
                        "pattern": "^\\+[1-9]\\d{7,14}$",
                        "examples": [
                            "+2250717738299"
                        ]
                    },
                    "channel": {
                        "$ref": "#/components/schemas/OtpChannel"
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "OtpRequestRequest"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}